Use minified bootstrap
I changed the Bootstrap reference in the head to use the minified
version. Minified files have all the extra white space and line breaks
removed. They're not meant to be readable, they're meant to load fast.
Since you shouldn't be editing the Bootstrap file, this is the
preferred way to work.